Refactor option classes and add support classes
